chore(Dockerfile): remove unused community modules from installation script
Remove the installation of unused community modules `mod_register`
from the Dockerfile. This change reduces the size and complexity of
the Docker image by excluding unnecessary modules, which are not
required for the current application setup. This helps in optimizing
the build process and potentially reduces security risks by limiting
the number of installed components.