feat(config): enable user registration and allow HTTP access from all interfaces
Enable user registration by setting `allow_registration` to true in the
core configuration. This change allows new users to register on the
server, which is essential for expanding the user base.

Modify the `http_interfaces` setting to allow HTTP access from all
interfaces by changing it from `127.0.0.1` to `*`. This change is
necessary to make the server accessible from external networks, which
is crucial for users to connect to the server from different devices
and locations.