chore(Dockerfile): improve Dockerfile structure and add hadolint ignore
Add a newline at the start for better readability. Include
'hadolint ignore' comments to suppress specific linter warnings
for apt-get usage. Refactor the Dockerfile to use WORKDIR for
better organization and readability. This change also combines
mkdir commands to reduce the number of layers in the final image.

These changes enhance the maintainability and readability of the
Dockerfile, making it easier to understand and modify in the
future. The use of WORKDIR improves the build context management,
and the hadolint ignore comments help in maintaining a clean
linter output.